tests/wpt/web-platform-tests/css/css-shapes/shape-outside-invalid-circle-000.html

<!DOCTYPE html>
<head>
    <meta charset="utf-8">
    <title>CSS Shapes Tests: Circle - invalid 2 args</title>
    <link rel="author" title="biqing" href="mailto:biqing.kwok@qq.com">
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes">
    <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#funcdef-circle">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <meta name="flags" content="dom">
    <meta name="assert" content="shape-outside declaration is invalid if circle shape function only has two parameters">
    <style type="text/css">
    #float {
        shape-outside: circle(50px, 50px);
        float: left;
        width: 100px;
        height: 100px;
    }
    </style>
</head>
<body>
    <div id="log"></div>

    <div id="float"></div>

    <script>

    test(function() {
        assert_true(getComputedStyle(document.getElementById('float')).shapeOutside == 'none');
    }, 'Circle - invalid 2 args');

    </script>
</body>
</html>
